run Eclipse code clean up

+++ src/main/java/org/apache/commons/net/TimeTCPClient.java
@@ -77,7 +77,7 @@ public final class TimeTCPClient extends SocketClient
     {
         DataInputStream input;
         input = new DataInputStream(_input_);
-        return (long)(input.readInt() & 0xffffffffL);
+        return input.readInt() & 0xffffffffL;
     }
 
     /***
